Allow specifying an optional cluster "association"
Where an association is how we tag the AWS resources.  This is so that the VPC
template can tag subnets with the same string as the cluster template itself.

For example, the VPC creation template may have a stack name of `my-stack`, and
it will instantiate the kubernetes-cluster template, causing another stack to
get created that looks like `my-stack-K8sStack-1234abcd`.  Since resources in
this sub-stack need to be tagged with the same thing as the VPC stack, we need
to support passing this as a string.

If the string isn't provided (or left blank), it falls back on using the
current stack name, which preserves previous behavior.

Signed-off-by: Ken Simon <ninkendo@gmail.com>