Configure for remote access
- Expose the API server over ELB
- Place a working kube config in the home directory for users to install to
  ~/.kube/config

This creates a load balancer inside the CFn template that routes to the
master instance on 6443.  The master adds itself using the `aws elb`
command on startup so that we don't have a dependency cycle.

To do this we're adding yet another (optional) parameter to the cluster
template: the public subnet ID.  If you leave it blank, it will be left
the same as the cluster subnet.  With that parameter added, over in the
VPC template we pass the just-created public subnet as that ID.

We need this because you cannot create a load balancer without
specifying the subnet or AZ id's, and using the AZ ID doesn't select the
right subnet (it doesn't know what VPC to use).  There doesn't appear to
be a way to assign VPC's directly on load balancers so that means we
need to specify it.

Signed-off-by: Ken Simon <ninkendo@gmail.com>